feat: add checkin command
